requirements.txt
channels==3.0.4
coreapi==2.3.3
coreschema==0.0.4
daphne==3.0.2
Django==3.2.12
django-colorfield==0.2.1
django-cors-headers==3.2.1
django-db-logger==0.1.10
django-debug-toolbar==2.2
django-filter==2.1.0
django-imagekit==4.0.2
django-log-viewer==1.1.2
django-model-utils==3.1.2
django-storages==1.12.3
django-health-check==3.16.5
djangorestframework==3.11.0
drf-api-logger==1.0.6
drf-history==1.0.1
drf-yasg==1.21.3
Pillow==8.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.3
python-memcached==1.59
requests==2.22.0
XlsxWriter==3.0.6
boto3==1.34.84
celery==5.1.2
redis==4.3.6
gunicorn
